feat: nouvelle version complète — Spring Boot 3 + React TypeScript + sub-path /gestion
- Backend : toutes les entités/controllers/services/repos de la nouvelle version
  (Inventaire, RGBI, RSR, Carburant, Bibliothèque, Commercial, StockEtats, Grade/Fonction/ServiceRef, Employe)
- Frontend : React TypeScript complet avec Vite, TailwindCSS, React Router
  base='/gestion' + BrowserRouter basename='/gestion' pour déploiement sub-path
  API_BASE='/gestion' en prod pour routage via nginx
- Docker : frontend sur port 8082:80 (main nginx proxie /gestion/)
- nginx.conf : proxy /api/ → backend:8080/api/

+++ backend/src/main/java/com/gestionmagasin/repository/FactureRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface FactureRepository extends JpaRepository<Facture, Integer> {
-    List<Facture> findByClientId(int clientId);
+public interface FactureRepository extends JpaRepository<Facture, Long> {
+    List<Facture> findByClientId(Long clientId);
     Optional<Facture> findByInvoiceNumber(String invoiceNumber);
 }
