+++ backend/src/main/java/com/gestionmagasin/repository/ItemRepository.java
@@ -4,6 +4,6 @@ import com.gestionmagasin.model.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ItemRepository extends JpaRepository<Item, Integer> {
-    List<Item> findByInvoiceId(int invoiceId);
+public interface ItemRepository extends JpaRepository<Item, Long> {
+    List<Item> findByInvoiceId(Long invoiceId);
 }
