+++ backend/src/main/java/com/gestionmagasin/repository/ProduitRepository.java
@@ -3,6 +3,6 @@ package com.gestionmagasin.repository;
 import com.gestionmagasin.model.Produit;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProduitRepository extends JpaRepository<Produit, Integer> {
+public interface ProduitRepository extends JpaRepository<Produit, Long> {
     Produit findByProducerCode(String producerCode);
 }
