+++ backend/src/main/java/com/gestionmagasin/repository/VendeurRepository.java
@@ -4,7 +4,7 @@ import com.gestionmagasin.model.Vendeur;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface VendeurRepository extends JpaRepository<Vendeur, Integer> {
+public interface VendeurRepository extends JpaRepository<Vendeur, Long> {
     Vendeur findByVendorCode(String vendorCode);
     List<Vendeur> findByVendorNameContainingIgnoreCase(String name);
 }
